Move tshirt_state to dedicated header file.
Creating tshirt_state.h and moving tshirt::tshirt_state struct from
Tshirt.h to this new file.
