+++ models/tshirt/include/Tshirt.h
@@ -8,35 +8,13 @@
 #include "Pdm03.h"
 #include "GIUH.hpp"
 #include "tshirt_params.h"
+#include "tshirt_state.h"
 #include <cmath>
 #include <utility>
 #include <vector>
 #include <memory>
 
 namespace tshirt {
-
-    //! Tshirt parameters struct
-
-    /*!
-     * Tshirt state structure
-     *
-     * This structure provides storage for the state used by Tshirt hydrological model at a particular time step.
-     */
-    struct tshirt_state {
-        // TODO: confirm this is correct
-        double soil_storage_meters;              //!< current water storage in soil column nonlinear reservoir ("Ss")
-        double groundwater_storage_meters;       //!< current water storage in ground water nonlinear reservoir ("Sgw")
-        vector<double> nash_cascade_storeage_meters;    //!< water storage in nonlinear reservoirs of Nash Cascade for lateral subsurface flow
-
-        // I think this doesn't belong in state, and so is just in run() below
-        //double column_total_soil_moisture_deficit;    //!< soil column total moisture deficit
-
-        tshirt_state(double soil_storage_meters = 0.0, double groundwater_storage_meters = 0.0,
-                     vector<double> nash_cascade_storeage_meters = vector<double>())
-                : soil_storage_meters(soil_storage_meters),
-                  groundwater_storage_meters(groundwater_storage_meters),
-                  nash_cascade_storeage_meters(std::move(nash_cascade_storeage_meters)) {}
-    };
 
     /*!
      * Tshirt flux structure
