+++ src/models/tshirt/Tshirt.cpp
@@ -1,4 +1,5 @@
 #include "Tshirt.h"
+#include "tshirt_state.h"
 #include "Constants.h"
 #include <cmath>
